foi validado o price, incluido search
