+++ CatalogoLivros/Service/IBookService.cs
@@ -6,6 +6,7 @@ namespace CatalogoLivros.Service
     {
         Task<IEnumerable<Book>> GetBooks();
         Task<Book> GetBookById(int id);
+        Task<IEnumerable<Book>> InsertBook(long isbn);
         Task<IEnumerable<Book>> GetBookByTitle(string title);
         Task CreateBook(Book book);
         Task UpdateBook(Book book);
