+++ React/catalogolivros/package.json
@@ -13,7 +13,9 @@
     "axios": "^1.2.1",
     "bootstrap": "^5.2.3",
     "react": "^18.2.0",
+    "react-bootstrap": "^2.7.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.5",
     "react-scripts": "5.0.1",
     "reactstrap": "^9.1.5",
     "typescript": "^4.9.4",
