+++ React/catalogolivros/src/App.tsx
@@ -1,303 +1,46 @@
-import React, { useState, useEffect } from "react";
-import './App.css';
-import axios from "axios";
+import Main from "./Main";
+import { Routes, Route } from 'react-router-dom';
 import "bootstrap/dist/css/bootstrap.min.css";
-import { Modal, ModalBody, ModalFooter, ModalHeader } from "reactstrap";
 
-function App() {
-  const baseUrl = "https://localhost:7043/api/Books";
-
-  const [data, setData] = useState([]);
-  const [modalIncluir, setModalIncluir] = useState(false);
-  const [modalEditar, setModalEditar] = useState(false);
-  const [modalExcluir, setModalExcluir] = useState(false);
-
-  const [bookSelected, setbookSelected] = useState({
-    id: "",
-    isbn: "",
-    title: "",
-    author: "",
-    price: "",
-  });
-
-  const selectBook = (book: any, opcao: string) => {
-    setbookSelected(book);
-    opcao === "Editar" ? abrirFecharModalEditar() : abrirFecharModalExcluir();
-  };
-
-  const abrirFecharModalIncluir = () => {
-    setModalIncluir(!modalIncluir);
-  };
-  const abrirFecharModalEditar = () => {
-    setModalEditar(!modalEditar);
-  };
-  const abrirFecharModalExcluir = () => {
-    setModalExcluir(!modalExcluir);
-  };
-
-  const handleChange = (e: any) => {
-    const { name, value } = e.target;
-    setbookSelected({
-      ...bookSelected,
-      [name]: value,
-    });
-    console.log(bookSelected);
-  };
-
-  const pedidoGet = async () => {
-    await axios
-      .get(baseUrl)
-      .then((response) => {
-        setData(response.data);
-      })
-      .catch((error) => {
-        console.log(error);
-      });
-  };
-
-  const pedidoPost = async () => {
-    delete bookSelected.id;
-    await axios
-      .post(baseUrl, bookSelected)
-      .then((response) => {
-        setData(data.concat(response.data));
-        abrirFecharModalIncluir();
-      })
-      .catch((error) => {
-        console.log(error);
-      });
-  };
-
-  const pedidoPut = async () => {
-    await axios
-      .put(baseUrl + "/" + bookSelected.id, bookSelected)
-      .then((response) => {
-        var resposta = response.data;
-        var dadosAuxiliar = data;
-        dadosAuxiliar.map((book: { id: string; isbn: number; title: string; author: string; price: number;}) => {
-          if (book.id === bookSelected.id) {
-            book.isbn = resposta.isbn;
-            book.title = resposta.title;
-            book.author = resposta.author;
-            book.price = resposta.price;
-          }
-        });
-        abrirFecharModalEditar();
-      })
-      .catch((error) => {
-        console.log(error);
-      });
-  };
-
-  const pedidoDelete = async () => {
-    await axios
-      .delete(baseUrl + "/" + bookSelected.id)
-      .then((response) => {
-        setData(data.filter((book) => book !== response.data));
-        abrirFecharModalExcluir();
-      })
-      .catch((error) => {
-        console.log(error);
-      });
-  };
-
-  useEffect(() => {
-    pedidoGet();
-  });
-
+export default function App() {
   return (
-    <div className="aluno-container">
-      <br />
-      <h3>Catálogo de Livros</h3>
-      <header>
-      
-        <button
-          className="btn btn-success"
-          onClick={() => abrirFecharModalIncluir()}
-        >
-          Incluir novo Aluno
-        </button>
-      </header>
-      <table className="table table-bordered">
-        <thead>
-          <tr>
-            <th scope="col">Id</th>
-            <th scope="col">Isbn</th>
-            <th scope="col">Title</th>
-            <th scope="col">Author</th>
-            <th scope="col">Price</th>
-            <th scope="col">Operação</th>
-          </tr>
-        </thead>
-        <tbody>
-          {data.map((book: { id: number; isbn: number; title: string; author: string; price : number; }) => (
-            <tr key={book.id}>
-              <td>{book.id}</td>
-              <td>{book.isbn}</td>
-              <td>{book.title}</td>
-              <td>{book.author}</td>
-              <td>{book.price}</td>
-              <td>
-              
-                <button
-                  className="btn btn-primary"
-                  onClick={() => selectBook(book, "Editar")}
-                >
-                  Editar
-                </button>{" "}
-                <button
-                  className="btn btn-danger"
-                  onClick={() => selectBook(book, "Excluir")}
-                >
-                  Excluir
-                </button>
-              </td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
+    <div>
+      <nav className="navbar navbar-expand-lg bg-primary">
+        <div className="container-fluid">
+          <a className="navbar-brand">
+            Catálogo de Livros
+          </a>
+          <button
+            className="navbar-toggler"
+            type="button"
+            data-bs-toggle="collapse"
+            data-bs-target="#navbarSupportedContent"
+            aria-controls="navbarSupportedContent"
+            aria-expanded="false"
+            aria-label="Toggle navigation"
+          >
+            <span className="navbar-toggler-icon"></span>
+          </button>
+          <div className="collapse navbar-collapse" id="navbarSupportedContent">
+            <ul className="navbar-nav me-auto mb-2 mb-lg-0">
+              <li className="nav-item">
+                <a className="nav-link active" aria-current="page" href="/">
+                  Home
+                </a>
+              </li>
+              <li className="nav-item">
+                <a className="nav-link" href="/books">
+                  Books
+                </a>
+              </li>
+            </ul>
 
-      {/* Modal incluir alunos */}
-      <Modal isOpen={modalIncluir}>
-        <ModalHeader>Incluir novo Aluno</ModalHeader>
-        <ModalBody>
-          <div className="form-group">
-            <label>Isbn: </label>
-            <br />
-            <input
-              type="text"
-              className="form-control"
-              name="isbn"
-              onChange={handleChange}
-            />
-            <br />
-            <label>Title: </label>
-            <br />
-            <input
-              type="text"
-              className="form-control"
-              name="title"
-              onChange={handleChange}
-            />
-            <br />
-            <label>Author: </label>
-            <input
-              type="text"
-              className="form-control"
-              name="author"
-              onChange={handleChange}
-            />
-            <br />
-            <label>price: </label>
-            <input
-              type="number"
-              className="form-control"
-              name="price"
-              onChange={handleChange}
-            />
-            <br />
           </div>
-        </ModalBody>
-        <ModalFooter>
-          <button className="btn btn-primary" onClick={() => pedidoPost()}>
-            Incluir
-          </button>{" "}
-          {"   "}
-          <button
-            className="btn btn-danger"
-            onClick={() => abrirFecharModalIncluir()}
-          >
-            Cancelar
-          </button>
-        </ModalFooter>
-      </Modal>
+        </div>
+      </nav>
 
-      {/* Modal Editar Alunos */}
-      <Modal isOpen={modalEditar}>
-        <ModalHeader>Editar Aluno</ModalHeader>
-        <ModalBody>
-          <div className="form-group">
-            <label>Id: </label>
-            <input
-              type="text"
-              className="form-control"
-              readOnly
-              value={bookSelected && bookSelected.id}
-            />
-            <br />
-            <label>Isbn: </label>
-            <br />
-            <input
-              type="text"
-              className="form-control"
-              name="isbn"
-              onChange={handleChange}
-              value={bookSelected && bookSelected.isbn}
-            />
-            <br />
-            <label>Title: </label>
-            <br />
-            <input
-              type="text"
-              className="form-control"
-              name="title"
-              onChange={handleChange}
-              value={bookSelected && bookSelected.title}
-            />
-            <br />
-            <label>Author: </label>
-            <input
-              type="text"
-              className="form-control"
-              name="author"
-              onChange={handleChange}
-              value={bookSelected && bookSelected.author}
-            />
-            <br />
-            <label>Price: </label>
-            <input
-              type="text"
-              className="form-control"
-              name="price "
-              onChange={handleChange}
-              value={bookSelected && bookSelected.price}
-            />
-            <br />
-          </div>
-        </ModalBody>
-        <ModalFooter>
-          <button className="btn btn-primary" onClick={() => pedidoPut()}>
-            Editar
-          </button>{" "}
-          {"   "}
-          <button
-            className="btn btn-danger"
-            onClick={() => abrirFecharModalEditar()}
-          >
-            Cancelar
-          </button>
-        </ModalFooter>
-      </Modal>
-
-      {/* Modal excluir alunos */}
-      <Modal isOpen={modalExcluir}>
-        <ModalBody>
-          Confirma a exclusão do livro {bookSelected && bookSelected.title} ?
-        </ModalBody>
-        <ModalFooter>
-          <button className="btn btn-danger" onClick={() => pedidoDelete()}>
-            Sim
-          </button>
-          <button
-            className="btn btn-secondary"
-            onClick={() => abrirFecharModalExcluir()}
-          >
-            Não
-          </button>
-        </ModalFooter>
-      </Modal>
+      <hr />
+      <Main />
     </div>
   );
 }
-
-export default App;
