+++ React/catalogolivros/src/pages/home.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import './homeStyle.css';
+
+export default function Home(){
+    return(
+        <h1>Home</h1>
+    );
+}
